Version 1.1
Cured the glitch that the centering is not centering when the end of the file is not as far as the bottom of the window.  Added a couple of new modes, these act just like zt and zb in normal mode. They are shift-ESC and shift-F1 respectively.  So now you've got all zz, zt & zb as smooth-equivalents: ESC, S-ESC and S-F1 respectively.