Version 1.4
Fixed a subtle bug that meant that if scrolloff>0 the display would keep on centering but not immediately. If you were at the top of bottom having done S-ESC or S-F1 with scrolloff>0 then the display would stop scrolling then, but then later on when you moved off that line the display would keep on trying to scroll. This was to do with where I had attempted to account for the 'scrolloff' value in the centering bit of my code.  Made the command line clear so that you don't see a silly 'call StartSmoothCentering()' at the bottom of the screen each time.