Version 1.3
Fixed the glitch where if you had a scrolloff>0 set it would carry on scrolling if you were at the top and bottom and you did a S-ESC or S-F1.  Now correctly identifies the scrolloff setting and accounts for this value, not just the cursors position within the current window.  Changed key mappings to be a more natural-feeling S-ESC gets you 'out of the code', in other words lets you see where you are by scrolling the file down, whereas S-F1 scrolls up and hence further deep into the file.