Version 1.6
Had a slight glitch that happened but not always when you did Shift-ESC, the scrolling would just continue. Not sure why but fixed it and checked it so that this does not happen.  Had a bug that caused the continuing of scrolling when line wrapping was on. This gave incorrect readings for the window height so that scrolling would continue.  Fixed this by putting in a check that warns you if wrap is on, and doesn't scroll.